+++ client/package.json
@@ -3,7 +3,11 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
+    "@lottiefiles/react-lottie-player": "^3.5.3",
+    "@react-pdf-viewer/core": "^3.12.0",
+    "@react-pdf-viewer/default-layout": "^3.12.0",
     "axios": "^1.5.0",
+    "pdfjs-dist": "^3.4.120",
     "react": "^18.2.0",
     "react-cookie": "^6.1.0",
     "react-cookies": "^0.1.1",
